All tasks

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -11,13 +11,13 @@
 */
 size_t list_len(const list_t *h)
 {
-   int count = 0;
+	size_t count = 0;
 
 
-   while (h)
-   {
-       count++;
-       h = h->next;
-   }
-   return (count);
+	while (h)
+	{
+		count++;
+		h = h->next;
+	}
+	return (count);
 }
